requirements.txt
langchain>=1.2
langchain-core>=1.2
langchain-openai>=1.1
huggingface_hub>=0.23
pandas>=2.0
streamlit>=1.35
python-dotenv>=1.0
kagglehub>=0.3
